Updated the linked data transformer.

+++ crm-jsonld-transformer-impl/src/main/java/ca/magex/crm/ld/common/TelephoneTransformer.java
@@ -22,8 +22,8 @@ public class TelephoneTransformer extends AbstractLinkedDataTransformer<Telephon
 	public Telephone parse(DataObject data) {
 		validateContext(data);
 		validateType(data);
-		Integer number = data.getInt("number");
-		Integer extension = data.getInt("extension");
+		Long number = data.getLong("number");
+		Long extension = data.getLong("extension");
 		return new Telephone(number, extension);
 	}
